requirements.txt
alembic==1.8.1
amqp==2.6.1
aniso8601==9.0.1
appdirs==1.4.4
async-timeout==4.0.2
attrs==21.4.0
bcrypt==3.2.2
billiard==3.6.4.0
black==22.6.0
celery==4.3.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.5
dataclasses==0.6
Deprecated==1.2.13
Flask==2.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-Migrate==3.1.0
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.8.3
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
jsonschema==3.2.0
kombu==4.6.11
Mako==1.2.1
MarkupSafe==2.0.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
Pillow==8.4.0
platformdirs==2.4.0
pycparser==2.21
PyJWT==2.4.0
PyMySQL==1.0.2
pyotp==2.6.0
pyparsing==3.0.9
pyrsistent==0.18.0
python-dotenv==0.20.0
pytz==2022.1
redis==4.3.3
regex==2022.4.24
requests==2.28.0
six==1.16.0
SQLAlchemy==1.4.36
stripe==4.0.2
tomli==1.2.3
typed-ast==1.5.4
typing_extensions==4.1.1
urllib3==1.26.9
vine==1.3.0
Werkzeug==2.0.3
wrapt==1.14.1
zipp==3.6.0
